tdd tests

+++ src/main/java/ISell.java
@@ -1,5 +1,5 @@
 public interface ISell {
 
-    public int calculateMarkup();
+    public double calculateMarkup();
 
 }
